[FEAT] 친구 위치의 날씨 정보 구현
- 친구 위치의 날씨 정보 받아오기 구현 완료

해야할 것
- 미세먼지 화면 디자인 수정
- 단기예보 리사이클러뷰 클릭 시 세부 예보 사항 보이게
- 위젯 혹은 알림기능 구현
