+++ app/src/main/res/layout/activity_weather.xml
@@ -163,7 +163,7 @@
                 <LinearLayout
                     android:id="@+id/ly_rv_very_short"
                     android:layout_width="match_parent"
-                    android:layout_height="250dp"
+                    android:layout_height="270dp"
                     android:background="@drawable/rounded_background"
                     android:orientation="vertical"
                     android:layout_marginTop="20dp">
