+++ app/src/main/res/layout/weather_toolbar.xml
@@ -20,11 +20,11 @@
             android:contentDescription="네비게이션 드로어 열기"
             android:scaleType="fitCenter"
             android:src="@drawable/menu"
-            android:tint="@color/white"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toStartOf="@+id/linearLayout3"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toTopOf="parent"
+            app:tint="@color/white" />
 
         <LinearLayout
             android:id="@+id/linearLayout3"
